api/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI & NLP
openai==1.3.0
anthropic==0.7.8
textblob==0.17.1
nltk==3.8.1
spacy==3.7.2

# HTTP & API
httpx==0.25.2
requests==2.31.0

# Data Processing
pandas==2.1.3
numpy==1.24.3

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
